Revert "Reorganized whisper implementation with separate OpenAI and HF modules"
