+++ psifx/audio/transcription/whisper/__init__.py
@@ -1,7 +1,1 @@
-"""whisper transcription module."""
-
-from psifx.audio.transcription.whisper.openai import OpenAIWhisperTool
-from psifx.audio.transcription.whisper.huggingface import HuggingFaceWhisperTool
-from psifx.audio.transcription.whisper.command import WhisperCommand
-
-__all__ = ["OpenAIWhisperTool", "HuggingFaceWhisperTool", "WhisperCommand"]
+"""Whisper transcription module."""
