--- psifx/audio/transcription/whisper/openai/__init__.py
@@ -1,6 +1,0 @@
-"""openai whisper module."""
-
-from psifx.audio.transcription.whisper.openai.tool import OpenAIWhisperTool
-from psifx.audio.transcription.whisper.openai.command import OpenAIWhisperCommand
-
-__all__ = ["OpenAIWhisperTool", "OpenAIWhisperCommand"]
