--- psifx/audio/transcription/whisper/openai/command.py
@@ -1,95 +1,0 @@
-"""OpenAI Whisper transcription command-line interface."""
-import argparse
-from pathlib import Path
-
-from psifx.utils.command import Command
-from psifx.audio.transcription.whisper.openai.tool import OpenAIWhisperTool
-
-
-class OpenAIWhisperCommand(Command):
-    """
-    Command-line interface for transcribing an audio track with OpenAI Whisper.
-    """
-
-    @staticmethod
-    def setup(parser: argparse.ArgumentParser):
-        """
-        Sets up the command.
-
-        :param parser: The argument parser.
-        :return:
-        """
-        parser.add_argument(
-            "--audio",
-            type=Path,
-            required=True,
-            help="path to the input audio file, such as ``/path/to/audio.wav``",
-        )
-        parser.add_argument(
-            "--transcription",
-            type=Path,
-            required=True,
-            help="path to the output transcription file, such as ``/path/to/transcription.vtt``",
-        )
-        parser.add_argument(
-            "--language",
-            type=str,
-            default=None,
-            help="language of the audio, if ignore, the model will try to guess it, "
-                 "it is advised to specify it",
-        )
-        parser.add_argument(
-            "--model_name",
-            type=str,
-            default="small",
-            help="name of the model, check https://github.com/openai/whisper#available-models-and-languages",
-        )
-        parser.add_argument(
-            "--translate_to_english",
-            default=False,
-            action=argparse.BooleanOptionalAction,
-            help="whether to transcribe the audio in its original language or"
-                 " to translate it to english",
-        )
-        parser.add_argument(
-            "--device",
-            type=str,
-            default="cpu",
-            help="device on which to run the inference, either 'cpu' or 'cuda'",
-        )
-        parser.add_argument(
-            "--overwrite",
-            default=False,
-            action=argparse.BooleanOptionalAction,
-            help="overwrite existing files, otherwise raises an error",
-        )
-        parser.add_argument(
-            "--verbose",
-            default=True,
-            action=argparse.BooleanOptionalAction,
-            help="verbosity of the script",
-        )
-
-    @staticmethod
-    def execute(parser: argparse.ArgumentParser, args: argparse.Namespace):
-        """
-        Executes the command.
-
-        :param parser: The argument parser.
-        :param args: The arguments.
-        :return:
-        """
-        tool = OpenAIWhisperTool(
-            model_name=args.model_name,
-            device=args.device,
-            overwrite=args.overwrite,
-            verbose=args.verbose,
-        )
-
-        tool.inference(
-            audio_path=args.audio,
-            transcription_path=args.transcription,
-            task="transcribe" if not args.translate_to_english else "translate",
-            language=args.language,
-        )
-        del tool
